scraper/requirements.txt
playwright==1.41.2
supabase==2.3.4
resend==0.7.1
python-dotenv==1.0.0
